Add comparison driver telemetry to Qualifying charts
A yellow line in now used to display a comparison driver's speed and gears telemetry in the Qualifying telemetry screen. The comparison lap is by default set to the pole position lap for that qualifying.

You can toggle the comparison telemetry by pressing the "C" key on your keyboard to turn it on and off.